+++ NEWS.md
@@ -8,6 +8,7 @@
     the end of the file
 -   **CAUTION**: `update_docs()` modifies RStudio Global and Project options.
 -   New capability for optimizing images with `resmush::resmush_dir()`.
+-   New function to check titles in `man` dir: `check_rd_titles()`.
 
 # pkgdev 0.1.0